getMADRankings implementation for rankable classes

+++ data/Ranking.java
@@ -2,18 +2,25 @@ package data;
 
 public class Ranking {
     public Team team;
-    public int points;
+    public double points;
 
     public Ranking(Team team, int points){
         this.team = team;
         this.points = points;
     }
 
+    /** 
+     * @return the team associated with the ranking
+     */
     public Team getTeam() {
         return team;
     }
 
-    public int getPoints() {
+    /** 
+     * @return return a number of points
+     * this amount of points is generic and can be for different statistics
+     */
+    public double getPoints() {
         return points;
     }
 }
